+++ src/app/components/home/home.component.ts
@@ -1,4 +1,5 @@
 import { Component, OnInit } from '@angular/core';
+import { FormGroup, FormBuilder } from '@angular/forms';
 import { ApiResponse } from '@app/interfaces/api-response';
 import { Party } from '@app/interfaces/party';
 import { ApiService } from '@app/services/api/api-service.service';
@@ -20,14 +21,35 @@ export class HomeComponent implements OnInit {
    */
   joinPartyId = '';
 
-  constructor(private api: ApiService) {
+  /**
+   * Filter parties form
+   */
+  exploreForm: FormGroup;
+
+  constructor(private api: ApiService,
+              private formBuilder: FormBuilder) {
   }
 
   ngOnInit(): void {
     /**
+     * Search form
+     */
+    this.exploreForm = this.formBuilder.group({
+      search: [''],
+    });
+    /**
      * Get parties
      */
-    this.api.getParties().subscribe((data: ApiResponse<Party>) => {
+    this.getParties();
+  }
+
+  /**
+   * Get parties for explore card
+   */
+  getParties(): void {
+    this.api.getParties({
+      search: this.exploreForm.get('search').value,
+    }).subscribe((data: ApiResponse<Party>) => {
       this.parties = data.results;
     });
   }