+++ src/app/services/api/api-service.service.ts
@@ -9,6 +9,7 @@ import { Party } from '@app/interfaces/party';
 import { PartyUser } from '@app/interfaces/party-user';
 import { Song } from '@app/interfaces/song';
 import { SongCategory } from '@app/interfaces/song-category';
+import { User } from '@app/interfaces/user';
 import { environment } from '@environments/environment';
 import { Observable } from 'rxjs';
 
@@ -28,6 +29,15 @@ export class ApiService {
   // User
 
   /**
+   * Get user data
+   *
+   * @param username User username
+   */
+  getUser(username: string): Observable<User> {
+    return this.http.get<User>(`${ApiService.base}users/${username}/`);
+  }
+
+  /**
    * Update user
    *
    * @param username User username
@@ -42,15 +52,35 @@ export class ApiService {
   /**
    * Get party list
    */
-  getParties(payload: { user?: number, status?: PartyStatus } = {}): Observable<ApiResponse<Party>> {
-    const params = new HttpParams();
-    if (payload.user) {
-      params.set('user', payload.user.toString());
-    }
-    if (payload.status) {
-      params.set('status', payload.status.toString());
-    }
+  getParties(payload: { user?: number, status?: PartyStatus, search?: string } = {}): Observable<ApiResponse<Party>> {
+    const filterPayload: {
+      user?: number,
+      status?: PartyStatus,
+      search?: string
+    } = this.filterObject<{ user?: number, status?: PartyStatus, search?: string }>(payload);
+    const params = new HttpParams({
+      fromObject: this.valuesToString(filterPayload),
+    });
     return this.http.get<ApiResponse<Party>>(`${ApiService.base}parties/`, { params });
+  }
+
+  // Convert to list
+  filterObject<T>(object: T): Partial<T> {
+    return Object.entries(object)
+      .filter(([key, value]): boolean => value !== null)
+      .reduce<Partial<T>>((accumulated: Partial<T>, [key, value]): Partial<T> => {
+        accumulated[key] = value;
+        return accumulated;
+      }, {});
+  }
+
+  // Convert to string
+  valuesToString<T>(object: T): { [P in keyof T]: string } {
+    return Object.entries(object)
+      .reduce<any>((accumulated: { [P in keyof T]: string }, [key, value]): { [P in keyof T]: string } => {
+        accumulated[key] = value.toString();
+        return accumulated;
+      }, {});
   }
 
   /**
@@ -106,10 +136,10 @@ export class ApiService {
    * Update category
    *
    * @param id Category ID
-   * @param name New category name
+   * @param payload New category name and image
    */
-  updateCategory(id: number, name: string): Observable<Category> {
-    return this.http.patch<Category>(`${ApiService.base}party-categories/${id}/`, { name });
+  updateCategory(id: number, payload: {name?: string, image?: string}): Observable<Category> {
+    return this.http.patch<Category>(`${ApiService.base}party-categories/${id}/`, payload);
   }
 
   /**
@@ -119,6 +149,15 @@ export class ApiService {
    */
   deleteCategory(id: number): Observable<void> {
     return this.http.delete<void>(`${ApiService.base}party-categories/${id}/`);
+  }
+
+  /**
+   * Get a party category
+   *
+   * @param id Category id
+   */
+  getCategory(id: number): Observable<Category> {
+    return this.http.get<Category>(`${ApiService.base}party-categories/${id}/`);
   }
 
   // Party user