+++ src/assets/changelog.md
@@ -2,8 +2,11 @@
 
 - Improve inputs to have labels instead of placeholders.
 - Improve player shuffle logic and design.
+- Add playlist image.
 - Add song to selected playlist.
-- Add link for categories (share link of categories)
+- Add link for categories (share link of categories).
+- Add search in exploring parties in dashboard.
+- Add profile page for users.
 - Fix issue when adding a new song.
 - Fix player popup is visible when player is expanded.